implemented gui

+++ src/main/model/Category.java
@@ -5,13 +5,15 @@ import java.util.HashMap;
 
 public class Category implements Serializable {
     private String category;
-    private HashMap<String, Integer> expenses;
+    private HashMap<String, Double> expenses;
 
     public Category(String category) {
         this.category = category;
         this.expenses = new HashMap<>();
     }
 
+    // MODIFIES: this
+    // EFFECTS: if Category doesn't already contain given Expenses e, sets Category of expense to this
     public void addExpense(Expenses e) {
         if (!expenses.containsKey(e.getType())) {
             expenses.put(e.getType(), e.getExpense());
@@ -19,7 +21,7 @@ public class Category implements Serializable {
         }
     }
 
-    public HashMap<String, Integer> getExpenses() {
+    public HashMap<String, Double> getExpenses() {
         return expenses;
     }
